Small Intestine automatic end-to-end suturing.
Details: fore most of cases, you don't need to suture two times an end-to-end small intestine. Also, when regions doesn't much we no more abort the suturing but correct the vector to get the right region. This way, we suture the selected region with the right matching region and then get the second region and suture it too (one ConnectTool process, two joints).